move spark 31 to spark310
