+++ shims/aggregator/pom.xml
@@ -21,12 +21,12 @@
 
     <parent>
         <groupId>com.nvidia</groupId>
-        <artifactId>rapids-4-spark-shims-aggregator_2.12</artifactId>
+        <artifactId>rapids-4-spark-shims_2.12</artifactId>
         <version>0.2.0-SNAPSHOT</version>
         <relativePath>../pom.xml</relativePath>
     </parent>
     <groupId>com.nvidia</groupId>
-    <artifactId>rapids-4-spark-shims_2.12</artifactId>
+    <artifactId>rapids-4-spark-shims-aggregator_2.12</artifactId>
     <packaging>jar</packaging>
     <name>RAPIDS Accelerator for Apache Spark SQL Plugin Shim Aggregator</name>
     <description>The RAPIDS SQL plugin for Apache Spark Shim Aggregator</description>
@@ -35,7 +35,7 @@
     <dependencies>
         <dependency>
             <groupId>com.nvidia</groupId>
-            <artifactId>rapids-4-spark-shims-spark31_${scala.binary.version}</artifactId>
+            <artifactId>rapids-4-spark-shims-spark310_${scala.binary.version}</artifactId>
             <version>${project.version}</version>
             <scope>compile</scope>
         </dependency>
@@ -47,7 +47,7 @@
         </dependency>
         <dependency>
             <groupId>com.nvidia</groupId>
-            <artifactId>rapids-4-spark-shims-spark30_${scala.binary.version}</artifactId>
+            <artifactId>rapids-4-spark-shims-spark300_${scala.binary.version}</artifactId>
             <version>${project.version}</version>
             <scope>compile</scope>
         </dependency>
