+++ shims/spark301/src/main/resources/META-INF/services/com.nvidia.spark.rapids.SparkShimServiceProvider
@@ -1,1 +1,1 @@
-com.nvidia.spark.rapids.shims.spark301.Spark301ShimServiceProvider
+com.nvidia.spark.rapids.shims.spark301.SparkShimServiceProvider
