--- shims/spark31/pom.xml
@@ -1,50 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-  Copyright (c) 2020, NVIDIA CORPORATION.
-
-  Licensed under the Apache License, Version 2.0 (the "License");
-  you may not use this file except in compliance with the License.
-  You may obtain a copy of the License at
-
-     http://www.apache.org/licenses/LICENSE-2.0
-
-  Unless required by applicable law or agreed to in writing, software
-  distributed under the License is distributed on an "AS IS" BASIS,
-  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  See the License for the specific language governing permissions and
-  limitations under the License.
--->
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-
-    <parent>
-        <groupId>com.nvidia</groupId>
-        <artifactId>rapids-4-spark-shims-aggregator_2.12</artifactId>
-        <version>0.2.0-SNAPSHOT</version>
-	<relativePath>../pom.xml</relativePath>
-    </parent>
-    <groupId>com.nvidia</groupId>
-    <artifactId>rapids-4-spark-shims-spark31_2.12</artifactId>
-    <name>RAPIDS Accelerator for Apache Spark SQL Plugin Spark 3.1 Shim</name>
-    <description>The RAPIDS SQL plugin for Apache Spark 3.1 Shim</description>
-    <version>0.2.0-SNAPSHOT</version>
-
-    <properties>
-        <spark31.version>3.1.0-SNAPSHOT</spark31.version>
-    </properties>
-
-    <dependencies>
-	<dependency>
-            <groupId>com.nvidia</groupId>
-            <artifactId>rapids-4-spark-shims-spark301_${scala.binary.version}</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>org.apache.spark</groupId>
-            <artifactId>spark-sql_${scala.binary.version}</artifactId>
-            <version>${spark31.version}</version>
-        </dependency>
-    </dependencies>
-</project>
