+++ shims/spark310/pom.xml
@@ -26,26 +26,26 @@
 	<relativePath>../pom.xml</relativePath>
     </parent>
     <groupId>com.nvidia</groupId>
-    <artifactId>rapids-4-spark-shims-spark301_2.12</artifactId>
-    <name>RAPIDS Accelerator for Apache Spark SQL Plugin Spark 3.0.1 Shim</name>
-    <description>The RAPIDS SQL plugin for Apache Spark 3.0.1 Shim</description>
+    <artifactId>rapids-4-spark-shims-spark310_2.12</artifactId>
+    <name>RAPIDS Accelerator for Apache Spark SQL Plugin Spark 3.1.0 Shim</name>
+    <description>The RAPIDS SQL plugin for Apache Spark 3.1.0 Shim</description>
     <version>0.2.0-SNAPSHOT</version>
 
     <properties>
-        <spark301.version>3.0.1</spark301.version>
+        <spark310.version>3.1.0-SNAPSHOT</spark310.version>
     </properties>
 
     <dependencies>
-        <dependency>
+	<dependency>
             <groupId>com.nvidia</groupId>
-            <artifactId>rapids-4-spark-shims-spark300_${scala.binary.version}</artifactId>
+            <artifactId>rapids-4-spark-shims-spark301_${scala.binary.version}</artifactId>
             <version>${project.version}</version>
         </dependency>
         <dependency>
             <groupId>org.apache.spark</groupId>
             <artifactId>spark-sql_${scala.binary.version}</artifactId>
-            <version>${spark301.version}</version>
-            <scope>provided</scope>
+            <version>${spark310.version}</version>
+	    <scope>provided</scope>
         </dependency>
     </dependencies>
 </project>
