+++ shims/spark310/src/main/scala/com/nvidia/spark/rapids/shims/spark310/SparkShimServiceProvider.scala
@@ -1,0 +1,32 @@
+/*
+ * Copyright (c) 2020, NVIDIA CORPORATION.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.nvidia.spark.rapids.shims.spark310
+
+import com.nvidia.spark.rapids.SparkShims
+
+class Spark31ShimServiceProvider extends com.nvidia.spark.rapids.SparkShimServiceProvider {
+
+  val VERSIONNAMES = Seq("3.1.0-SNAPSHOT", "3.1.0")
+
+  def matchesVersion(version: String): Boolean = {
+    VERSIONNAMES.contains(version)
+  }
+
+  def buildShim: SparkShims = {
+    new Spark310Shims()
+  }
+}
